# perf: add multiple optimizations