--- a/include/ada/url_aggregator-inl.h
+++ b/include/ada/url_aggregator-inl.h
@@ -118,10 +118,9 @@ inline void url_aggregator::update_base_hostname(const std::string_view input) {
 
 ada_really_inline uint32_t url_aggregator::get_pathname_length() const noexcept {
   ada_log("url_aggregator::get_pathname_length");
-  uint32_t ending_index = 0;
+  uint32_t ending_index = uint32_t(buffer.size());
   if (components.search_start != url_components::omitted) { ending_index = components.search_start; }
   else if (components.hash_start != url_components::omitted) { ending_index = components.hash_start; }
-  else { ending_index = uint32_t(buffer.size()); }
   return ending_index - components.pathname_start;
 }