Running: `sudo build/benchmarks/bench ./datasets/wikipedia/wikipedia_100k.txt`

Before

```
BasicBench_AdaURL_aggregator              21961383 ns     21959687 ns           32 GHz=3.21847 cycle/byte=13.6115 cycles/url=695.45 instructions/byte=62.6634 instructions/cycle=4.6037 instructions/ns=14.8169 instructions/url=3.20164k ns/url=216.081 speed=232.666M/s time/byte=4.29801ns time/url=219.597ns url/s=4.5538M/s
BasicBench_AdaURL_aggregator_just_parse   22256599 ns     22252937 ns           32 GHz=3.22738 cycle/byte=13.73 cycles/url=701.5 instructions/byte=63.6907 instructions/cycle=4.63881 instructions/ns=14.9712 instructions/url=3.25412k ns/url=217.359 speed=229.6M/s time/byte=4.35541ns time/url=222.529ns url/s=4.49379M/s
```

After

```
----------------------------------------------------------------------------------------------------
Benchmark                                        Time             CPU   Iterations instructions/byte
----------------------------------------------------------------------------------------------------
BasicBench_AdaURL                         20176538 ns     20176514 ns           35           57.0265
BasicBench_AdaURL_just_parse              12578289 ns     12575679 ns           56           35.6931
BasicBench_AdaURL_aggregator              20621221 ns     20616647 ns           34           55.7308
BasicBench_AdaURL_aggregator_just_parse   20411587 ns     20410382 ns           34           55.5922
```